Moved AccessToken from sample-app to azure-core. Renamed storage.blob.credentials to storage.blob.credential. Made slight corrections to the TokenRequest classes' Javadoc in sample-app.
