+++ samples/sample-app-storage/src/main/java/com/azure/android/storage/sample/MsalClient.java
@@ -5,14 +5,13 @@ import android.util.Log;
 
 import androidx.annotation.NonNull;
 
-import com.azure.android.storage.sample.core.credential.AccessToken;
+import com.azure.android.core.credential.AccessToken;
 import com.azure.android.storage.sample.core.util.tokenrequest.TokenResponseCallback;
 import com.microsoft.identity.client.AuthenticationCallback;
 import com.microsoft.identity.client.IAccount;
 import com.microsoft.identity.client.IAuthenticationResult;
 import com.microsoft.identity.client.IMultipleAccountPublicClientApplication;
 import com.microsoft.identity.client.IPublicClientApplication;
-import com.microsoft.identity.client.PublicClientApplication;
 import com.microsoft.identity.client.SilentAuthenticationCallback;
 import com.microsoft.identity.client.exception.MsalClientException;
 import com.microsoft.identity.client.exception.MsalException;
