+++ samples/sample-app-storage/src/main/java/com/azure/android/storage/sample/core/util/tokenrequest/TokenResponseCallback.java
@@ -1,22 +1,22 @@
 package com.azure.android.storage.sample.core.util.tokenrequest;
 
-import com.azure.android.storage.sample.core.credential.AccessToken;
+import com.azure.android.core.credential.AccessToken;
 
 /**
  * The callback that {@link TokenRequestObserver} uses to notify the result of token retrieval.
  */
 public interface TokenResponseCallback {
     /**
-     * Notify successful retrieval of access-token.
+     * Notify the successful retrieval of an access token.
      *
-     * @param token the retrieved token
+     * @param token The retrieved token.
      */
     void onToken(AccessToken token);
 
     /**
-     * Notify the error occurred during access-token retrieval.
+     * Notify the error occurred during access token retrieval.
      *
-     * @param t the error
+     * @param t The error.
      */
     void onError(Throwable t);
 }
